Move orchestration to constants
