+++ src/debugging/DockerPlatformHelper.ts
@@ -5,8 +5,6 @@
 
 export type DockerPlatform = 'netCore' | 'node';
 
-export type DockerOrchestration = 'single' | 'docker-compose';
-
 interface DockerPlatformConfiguration {
     platform?: DockerPlatform;
     netCore?: {};
